workspace/scripts/requirements.txt
google-cloud-bigquery>=3.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
